refactored 3 routes change_username, change_password, change_email
